Create pathways with Acetaminophen
This query asks WikiPathways to search for all the pathways that contain a data node with the word “Acetaminophen”, which is paracetamol. Using the SELECT clause it asks for different variables that will be defined by the query.  
In the query, it is defined what a data node is and defines the search for the data node Acetaminophen. 
The next part defines the variables pathway, pathwayID and the description. It also explains that the found pathways have to be those of humans.