Create genes that influence Acetaminophen
This query asks WikiPathways to search for all the genes that play a role in pathways WP4269 and WP5444. It asks to show the pathway title in which the gene is present and asks for the name of the gene. It also asks to show these results ordered starting with all the genes in pathway WP4269 and thereafter all the genes in pathway WP5444.