Create references used in pathways
This query is used to search for the used references that were noted in the pathway. This query focusses on pathways WP4269 and WP5444 in order to find the used references. this query what the results to tell the ID of the 2 pathways and also asks for the link to the used reference per pathway ID.